+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Breaking change: Support only Ruby 3.0 or later.
 
-- Add support Vue script setup.
+- Add support Vue script setup and lang.
 
 ## [0.3.15] - 2022-08-21